+++ tool/web_search.go
@@ -20,23 +20,14 @@ import (
 	"encoding/json"
 	"fmt"
 	stdhtml "html"
-	"image"
-	_ "image/gif"
-	_ "image/jpeg"
-	_ "image/png"
 	"io"
 	"net/http"
 	"net/url"
-	"os"
-	"path/filepath"
-	"regexp"
-	"strconv"
 	"strings"
 	"time"
 
 	"github.com/ThinkInAIXYZ/go-mcp/protocol"
 	"github.com/the-open-agent/openagent/proxy"
-	_ "golang.org/x/image/webp"
 	"golang.org/x/net/html"
 )
 
@@ -87,8 +78,6 @@ func (p *WebSearchTool) BuiltinTools() []BuiltinTool {
 		searchEngineID: p.searchEngineID,
 		endpoint:       p.endpoint,
 		httpClient:     p.httpClient,
-	}, &imageDownloadBuiltin{
-		httpClient: p.httpClient,
 	}}
 }
 
@@ -106,8 +95,6 @@ const (
 	webSearchTimeout         = 20 * time.Second
 	webSearchMaxResponseSize = 2 * 1024 * 1024
 	webSearchUserAgent       = "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/122.0.0.0 Safari/537.36"
-	imageSearchMaxImageSize  = 5 * 1024 * 1024
-	imageDownloadMaxSize     = 25 * 1024 * 1024
 )
 
 type webSearchEngine string
@@ -122,10 +109,7 @@ const (
 var (
 	webSearchHTTPClient          = &http.Client{Timeout: webSearchTimeout}
 	duckDuckGoHTMLSearchEndpoint = "https://html.duckduckgo.com/html"
-	duckDuckGoHomeEndpoint       = "https://duckduckgo.com/"
-	duckDuckGoImageEndpoint      = "https://duckduckgo.com/i.js"
 	bingHTMLSearchEndpoint       = "https://www.bing.com/search"
-	bingImageSearchEndpoint      = "https://www.bing.com/images/search"
 	googleJSONSearchEndpoint     = "https://www.googleapis.com/customsearch/v1"
 	baiduWebSearchEndpoint       = "https://qianfan.baidubce.com/v2/ai_search/web_search"
 )
@@ -175,53 +159,6 @@ type googleSearchResponse struct {
 	} `json:"error,omitempty"`
 }
 
-type imageSearchBuiltin struct {
-	engine         webSearchEngine
-	apiKey         string
-	searchEngineID string
-	endpoint       string
-	httpClient     *http.Client
-}
-
-type imageDownloadBuiltin struct {
-	httpClient *http.Client
-}
-
-type modelVisionContextKey struct{}
-
-func WithModelVision(ctx context.Context, enabled bool) context.Context {
-	return context.WithValue(ctx, modelVisionContextKey{}, enabled)
-}
-
-type imageSearchResult struct {
-	ID           string `json:"id"`
-	Title        string `json:"title,omitempty"`
-	ImageURL     string `json:"imageUrl"`
-	ThumbnailURL string `json:"thumbnailUrl,omitempty"`
-	SourceURL    string `json:"sourceUrl,omitempty"`
-	Width        int    `json:"width,omitempty"`
-	Height       int    `json:"height,omitempty"`
-}
-
-type imageSearchPayload struct {
-	Query           string                   `json:"query"`
-	Provider        string                   `json:"provider"`
-	Count           int                      `json:"count"`
-	ExternalContent webSearchExternalContent `json:"externalContent"`
-	Results         []imageSearchResult      `json:"results"`
-}
-
-type duckDuckGoImageResponse struct {
-	Results []struct {
-		Title     string `json:"title"`
-		Image     string `json:"image"`
-		Thumbnail string `json:"thumbnail"`
-		URL       string `json:"url"`
-		Width     int    `json:"width"`
-		Height    int    `json:"height"`
-	} `json:"results"`
-}
-
 type baiduWebSearchRequest struct {
 	Messages           []baiduWebSearchMessage      `json:"messages"`
 	SearchSource       string                       `json:"search_source"`
@@ -258,166 +195,6 @@ type baiduWebSearchResponse struct {
 
 func (w *webSearchBuiltin) GetName() string {
 	return "web_search"
-}
-
-func (t *imageSearchBuiltin) GetName() string {
-	return "image_search"
-}
-
-func (t *imageSearchBuiltin) GetDescription() string {
-	return `Search the web for images and inspect the returned thumbnails. Use this when visual comparison is needed before choosing an image. Returns image URLs, source pages, dimensions, and the actual thumbnails for visual analysis.`
-}
-
-func (t *imageSearchBuiltin) GetInputSchema() interface{} {
-	return (&webSearchBuiltin{}).GetInputSchema()
-}
-
-func (t *imageSearchBuiltin) Execute(ctx context.Context, arguments map[string]interface{}) (*protocol.CallToolResult, error) {
-	if isVision, ok := ctx.Value(modelVisionContextKey{}).(bool); ok && !isVision {
-		return webSearchToolError("当前模型不支持看图"), nil
-	}
-
-	params, err := parseWebSearchArguments(arguments)
-	if err != nil {
-		return webSearchToolError(err.Error()), nil
-	}
-
-	results, provider, err := t.runImageSearch(ctx, params)
-	if err != nil {
-		return webSearchToolError(fmt.Sprintf("Image search failed: %s", err.Error())), nil
-	}
-
-	content := make([]protocol.Content, 0, len(results)+1)
-	visibleResults := make([]imageSearchResult, 0, len(results))
-	for _, result := range results {
-		imageURL := result.ThumbnailURL
-		if imageURL == "" {
-			imageURL = result.ImageURL
-		}
-		data, mimeType, width, height, err := downloadImage(ctx, imageURL, imageSearchMaxImageSize, t.httpClient)
-		if err != nil {
-			continue
-		}
-		if result.Width == 0 {
-			result.Width = width
-		}
-		if result.Height == 0 {
-			result.Height = height
-		}
-		result.ID = fmt.Sprintf("image_%d", len(visibleResults)+1)
-		visibleResults = append(visibleResults, result)
-		content = append(content, &protocol.ImageContent{
-			Type:     "image",
-			Data:     data,
-			MimeType: mimeType,
-		})
-	}
-	if len(visibleResults) == 0 {
-		return webSearchToolError("Image search failed: no image thumbnails could be downloaded"), nil
-	}
-
-	payload := imageSearchPayload{
-		Query:    params.Query,
-		Provider: provider,
-		Count:    len(visibleResults),
-		ExternalContent: webSearchExternalContent{
-			Untrusted: true,
-			Source:    "image_search",
-		},
-		Results: visibleResults,
-	}
-	payloadBytes, err := json.Marshal(payload)
-	if err != nil {
-		return nil, err
-	}
-	content = append([]protocol.Content{
-		&protocol.TextContent{Type: "text", Text: string(payloadBytes)},
-	}, content...)
-	return &protocol.CallToolResult{Content: content}, nil
-}
-
-func (t *imageSearchBuiltin) runImageSearch(ctx context.Context, params webSearchParams) ([]imageSearchResult, string, error) {
-	switch t.engine {
-	case webSearchEngineGoogle:
-		results, err := runGoogleImageSearch(ctx, params, t.apiKey, t.searchEngineID, t.endpoint, t.httpClient)
-		return results, "google", err
-	case webSearchEngineBing:
-		results, err := runBingImageSearch(ctx, params, t.httpClient)
-		return results, "bing", err
-	case webSearchEngineDuckDuckGo:
-		results, err := runDuckDuckGoImageSearch(ctx, params, t.httpClient)
-		return results, "duckduckgo", err
-	case webSearchEngineBaidu:
-		results, err := runBaiduImageSearch(ctx, params, t.apiKey, t.endpoint, t.httpClient)
-		return results, "baidu", err
-	default:
-		return nil, "", fmt.Errorf("image search is not supported by %s", t.engine)
-	}
-}
-
-func (t *imageDownloadBuiltin) GetName() string {
-	return "image_download"
-}
-
-func (t *imageDownloadBuiltin) GetDescription() string {
-	return `Download a selected HTTP(S) image to a local path. Use an imageUrl returned by image_search.`
-}
-
-func (t *imageDownloadBuiltin) GetInputSchema() interface{} {
-	return map[string]interface{}{
-		"type":                 "object",
-		"additionalProperties": false,
-		"properties": map[string]interface{}{
-			"url": map[string]interface{}{
-				"type":        "string",
-				"description": "Direct HTTP(S) image URL.",
-			},
-			"path": map[string]interface{}{
-				"type":        "string",
-				"description": "Local destination path.",
-			},
-		},
-		"required": []string{"url", "path"},
-	}
-}
-
-func (t *imageDownloadBuiltin) Execute(ctx context.Context, arguments map[string]interface{}) (*protocol.CallToolResult, error) {
-	imageURL := readWebSearchString(arguments, "url", "")
-	outputPath := readWebSearchString(arguments, "path", "")
-	if imageURL == "" {
-		return webSearchToolError("missing required parameter: url"), nil
-	}
-	if outputPath == "" {
-		return webSearchToolError("missing required parameter: path"), nil
-	}
-	parsed, err := url.Parse(imageURL)
-	if err != nil || (parsed.Scheme != "http" && parsed.Scheme != "https") {
-		return webSearchToolError("url must use HTTP(S)"), nil
-	}
-
-	data, mimeType, width, height, err := downloadImage(ctx, imageURL, imageDownloadMaxSize, t.httpClient)
-	if err != nil {
-		return webSearchToolError(fmt.Sprintf("Image download failed: %s", err.Error())), nil
-	}
-	outputPath, err = filepath.Abs(outputPath)
-	if err != nil {
-		return webSearchToolError(fmt.Sprintf("Image download failed: %s", err.Error())), nil
-	}
-	if err := os.MkdirAll(filepath.Dir(outputPath), 0o755); err != nil {
-		return webSearchToolError(fmt.Sprintf("Image download failed: %s", err.Error())), nil
-	}
-	if err := os.WriteFile(outputPath, data, 0o644); err != nil {
-		return webSearchToolError(fmt.Sprintf("Image download failed: %s", err.Error())), nil
-	}
-
-	payload, _ := json.Marshal(map[string]interface{}{
-		"path":     outputPath,
-		"mimeType": mimeType,
-		"width":    width,
-		"height":   height,
-		"bytes":    len(data),
-	})
-	return webSearchToolText(string(payload), false), nil
 }
 
 func (t *webSearchBuiltin) GetDescription() string {
@@ -673,149 +450,6 @@ func runGoogleSearch(ctx context.Context, params webSearchParams, apiKey string,
 	return limitWebSearchResults(results, params.Count), nil
 }
 
-func runGoogleImageSearch(ctx context.Context, params webSearchParams, apiKey string, searchEngineID string, endpoint string, httpClient *http.Client) ([]imageSearchResult, error) {
-	if strings.TrimSpace(apiKey) == "" {
-		return nil, fmt.Errorf("Google search requires an API key in clientSecret")
-	}
-	if strings.TrimSpace(searchEngineID) == "" {
-		return nil, fmt.Errorf("Google search requires a search engine ID (cx) in clientId")
-	}
-
-	query := url.Values{}
-	query.Set("key", apiKey)
-	query.Set("cx", searchEngineID)
-	query.Set("q", params.Query)
-	query.Set("num", strconv.Itoa(params.Count))
-	query.Set("searchType", "image")
-	query.Set("safe", "active")
-	if params.Language != "" {
-		query.Set("hl", params.Language)
-	}
-	if params.Country != "" {
-		query.Set("gl", params.Country)
-	}
-
-	body, err := fetchWebSearchAPI(ctx, http.MethodGet, resolveWebSearchEndpoint(endpoint, googleJSONSearchEndpoint), query, nil, nil, httpClient)
-	if err != nil {
-		return nil, err
-	}
-	var response googleSearchResponse
-	if err := json.Unmarshal(body, &response); err != nil {
-		return nil, err
-	}
-	if response.Error != nil && response.Error.Message != "" {
-		return nil, fmt.Errorf("Google returned an error: %s", response.Error.Message)
-	}
-
-	results := make([]imageSearchResult, 0, len(response.Items))
-	for _, item := range response.Items {
-		if strings.TrimSpace(item.Link) == "" {
-			continue
-		}
-		results = append(results, imageSearchResult{
-			Title:        cleanWebSearchText(item.Title),
-			ImageURL:     strings.TrimSpace(item.Link),
-			ThumbnailURL: strings.TrimSpace(item.Image.ThumbnailLink),
-			SourceURL:    strings.TrimSpace(item.Image.ContextLink),
-			Width:        item.Image.Width,
-			Height:       item.Image.Height,
-		})
-	}
-	if len(results) == 0 {
-		return nil, fmt.Errorf("Google returned no image results")
-	}
-	return limitImageSearchResults(results, params.Count), nil
-}
-
-func runBingImageSearch(ctx context.Context, params webSearchParams, httpClient *http.Client) ([]imageSearchResult, error) {
-	query := url.Values{}
-	query.Set("q", params.Query)
-	query.Set("count", strconv.Itoa(params.Count))
-	query.Set("safeSearch", "Strict")
-	if params.Language != "" {
-		query.Set("setlang", params.Language)
-	}
-	if params.Country != "" {
-		query.Set("cc", params.Country)
-	}
-	body, err := fetchWebSearchHTML(ctx, bingImageSearchEndpoint, query, httpClient)
-	if err != nil {
-		return nil, err
-	}
-	results, err := parseBingImageHTML(body)
-	if err != nil {
-		return nil, err
-	}
-	if len(results) == 0 {
-		return nil, fmt.Errorf("Bing returned no image results")
-	}
-	return limitImageSearchResults(results, params.Count), nil
-}
-
-func runDuckDuckGoImageSearch(ctx context.Context, params webSearchParams, httpClient *http.Client) ([]imageSearchResult, error) {
-	form := url.Values{"q": []string{params.Query}}
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, duckDuckGoHomeEndpoint, strings.NewReader(form.Encode()))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("User-Agent", webSearchUserAgent)
-	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-	resp, err := httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	body, err := io.ReadAll(io.LimitReader(resp.Body, webSearchMaxResponseSize+1))
-	if err != nil {
-		return nil, err
-	}
-	if len(body) > webSearchMaxResponseSize {
-		return nil, fmt.Errorf("response body exceeds %d bytes", webSearchMaxResponseSize)
-	}
-	tokenMatch := regexp.MustCompile(`vqd=["']?([0-9-]+)`).FindSubmatch(body)
-	if len(tokenMatch) < 2 {
-		return nil, fmt.Errorf("DuckDuckGo image token was not found")
-	}
-
-	query := url.Values{}
-	query.Set("q", params.Query)
-	query.Set("vqd", string(tokenMatch[1]))
-	query.Set("o", "json")
-	query.Set("f", ",,,")
-	query.Set("p", "1")
-	if params.Country != "" && params.Language != "" {
-		query.Set("l", fmt.Sprintf("%s-%s", params.Country, params.Language))
-	}
-	data, err := fetchWebSearchAPI(ctx, http.MethodGet, duckDuckGoImageEndpoint, query, nil, map[string]string{
-		"Referer": duckDuckGoHomeEndpoint,
-	}, httpClient)
-	if err != nil {
-		return nil, err
-	}
-	var response duckDuckGoImageResponse
-	if err := json.Unmarshal(data, &response); err != nil {
-		return nil, err
-	}
-	results := make([]imageSearchResult, 0, len(response.Results))
-	for _, item := range response.Results {
-		if strings.TrimSpace(item.Image) == "" {
-			continue
-		}
-		results = append(results, imageSearchResult{
-			Title:        cleanWebSearchText(item.Title),
-			ImageURL:     strings.TrimSpace(item.Image),
-			ThumbnailURL: strings.TrimSpace(item.Thumbnail),
-			SourceURL:    strings.TrimSpace(item.URL),
-			Width:        item.Width,
-			Height:       item.Height,
-		})
-	}
-	if len(results) == 0 {
-		return nil, fmt.Errorf("DuckDuckGo returned no image results")
-	}
-	return limitImageSearchResults(results, params.Count), nil
-}
-
 func runBaiduSearch(ctx context.Context, params webSearchParams, apiKey string, endpoint string, httpClient *http.Client) ([]webSearchResult, error) {
 	response, err := fetchBaiduSearch(ctx, params, apiKey, endpoint, "web", httpClient)
 	if err != nil {
@@ -826,18 +460,6 @@ func runBaiduSearch(ctx context.Context, params webSearchParams, apiKey string, 
 		return nil, fmt.Errorf("Baidu returned no results")
 	}
 	return limitWebSearchResults(results, params.Count), nil
-}
-
-func runBaiduImageSearch(ctx context.Context, params webSearchParams, apiKey string, endpoint string, httpClient *http.Client) ([]imageSearchResult, error) {
-	response, err := fetchBaiduSearch(ctx, params, apiKey, endpoint, "image", httpClient)
-	if err != nil {
-		return nil, err
-	}
-	results := parseBaiduImageSearchResponse(response)
-	if len(results) == 0 {
-		return nil, fmt.Errorf("Baidu returned no image results")
-	}
-	return limitImageSearchResults(results, params.Count), nil
 }
 
 func fetchBaiduSearch(ctx context.Context, params webSearchParams, apiKey string, endpoint string, resourceType string, httpClient *http.Client) (baiduWebSearchResponse, error) {
@@ -1068,40 +690,6 @@ func parseBingHTML(body string) ([]webSearchResult, error) {
 	return results, nil
 }
 
-func parseBingImageHTML(body string) ([]imageSearchResult, error) {
-	root, err := html.Parse(strings.NewReader(body))
-	if err != nil {
-		return nil, err
-	}
-	nodes := findHTMLNodes(root, func(n *html.Node) bool {
-		return n.Type == html.ElementNode && n.Data == "a" && htmlNodeHasClass(n, "iusc")
-	})
-	results := make([]imageSearchResult, 0, len(nodes))
-	for _, node := range nodes {
-		var metadata struct {
-			Title        string `json:"t"`
-			ImageURL     string `json:"murl"`
-			ThumbnailURL string `json:"turl"`
-			SourceURL    string `json:"purl"`
-			Width        int    `json:"mw"`
-			Height       int    `json:"mh"`
-		}
-		raw := stdhtml.UnescapeString(htmlAttribute(node, "m"))
-		if raw == "" || json.Unmarshal([]byte(raw), &metadata) != nil || metadata.ImageURL == "" {
-			continue
-		}
-		results = append(results, imageSearchResult{
-			Title:        cleanWebSearchText(metadata.Title),
-			ImageURL:     metadata.ImageURL,
-			ThumbnailURL: metadata.ThumbnailURL,
-			SourceURL:    metadata.SourceURL,
-			Width:        metadata.Width,
-			Height:       metadata.Height,
-		})
-	}
-	return results, nil
-}
-
 func parseGoogleSearchResponse(response googleSearchResponse) []webSearchResult {
 	results := make([]webSearchResult, 0, len(response.Items))
 	for _, item := range response.Items {
@@ -1146,29 +734,6 @@ func parseBaiduSearchResponse(response baiduWebSearchResponse) []webSearchResult
 			URL:      resultURL,
 			Snippet:  cleanWebSearchText(reference.Content),
 			SiteName: siteName,
-		})
-	}
-	return results
-}
-
-func parseBaiduImageSearchResponse(response baiduWebSearchResponse) []imageSearchResult {
-	results := make([]imageSearchResult, 0, len(response.References))
-	for _, reference := range response.References {
-		if reference.Image == nil || strings.TrimSpace(reference.Image.URL) == "" {
-			continue
-		}
-		title := cleanWebSearchText(reference.Title)
-		if title == "" {
-			title = cleanWebSearchText(reference.WebAnchor)
-		}
-		width, _ := strconv.Atoi(reference.Image.Width)
-		height, _ := strconv.Atoi(reference.Image.Height)
-		results = append(results, imageSearchResult{
-			Title:     title,
-			ImageURL:  strings.TrimSpace(reference.Image.URL),
-			SourceURL: strings.TrimSpace(reference.URL),
-			Width:     width,
-			Height:    height,
 		})
 	}
 	return results
@@ -1304,58 +869,6 @@ func limitWebSearchResults(results []webSearchResult, count int) []webSearchResu
 		return results
 	}
 	return results[:count]
-}
-
-func limitImageSearchResults(results []imageSearchResult, count int) []imageSearchResult {
-	if len(results) <= count {
-		return results
-	}
-	return results[:count]
-}
-
-func downloadImage(ctx context.Context, rawURL string, limit int64, httpClient *http.Client) ([]byte, string, int, int, error) {
-	parsed, err := url.Parse(strings.TrimSpace(rawURL))
-	if err != nil || (parsed.Scheme != "http" && parsed.Scheme != "https") {
-		return nil, "", 0, 0, fmt.Errorf("invalid HTTP(S) image URL")
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, parsed.String(), nil)
-	if err != nil {
-		return nil, "", 0, 0, err
-	}
-	req.Header.Set("User-Agent", webSearchUserAgent)
-	req.Header.Set("Accept", "image/*")
-	resp, err := httpClient.Do(req)
-	if err != nil {
-		return nil, "", 0, 0, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode < http.StatusOK || resp.StatusCode >= http.StatusMultipleChoices {
-		return nil, "", 0, 0, fmt.Errorf("HTTP %d", resp.StatusCode)
-	}
-	if resp.ContentLength > limit {
-		return nil, "", 0, 0, fmt.Errorf("image exceeds %d bytes", limit)
-	}
-	data, err := io.ReadAll(io.LimitReader(resp.Body, limit+1))
-	if err != nil {
-		return nil, "", 0, 0, err
-	}
-	if int64(len(data)) > limit {
-		return nil, "", 0, 0, fmt.Errorf("image exceeds %d bytes", limit)
-	}
-	config, format, err := image.DecodeConfig(bytes.NewReader(data))
-	if err != nil {
-		return nil, "", 0, 0, fmt.Errorf("cannot decode image: %w", err)
-	}
-	mimeType := map[string]string{
-		"jpeg": "image/jpeg",
-		"png":  "image/png",
-		"gif":  "image/gif",
-		"webp": "image/webp",
-	}[format]
-	if mimeType == "" {
-		return nil, "", 0, 0, fmt.Errorf("unsupported image format %q", format)
-	}
-	return data, mimeType, config.Width, config.Height, nil
 }
 
 func resolveWebSearchSiteName(rawURL string) string {
